Adjust sign of piece values when plotting
Since gene scalars are allowed to go negative, the scalar values of the
Total Force Gene and the Opponent Pieces Targeted Gene can compensate
for negatively valued pieces by going negative. To get a consistent
picture of the worth of a piece across generations and gene pools,
multiply the value of a piece by the sign of the Total Force Gene.

This behavior doesn't make sense. The piece values returned by
Piece_Strength_Gene::piece_value() are normalized so that the total
starting value of the board is +1.0 for a side. If the piece values take
on negative values, then the normalization factor will be negative,
resulting in a positive return value. The signs of the piece values
inside the gene should not correlate with the Total Strength Gene and
the Opponent Pieces Targeted Gene scalars. And yet they do.

It looks like some AI players are pessimizing their piece strength
values. The last time this happened, the look ahead recursion had an
error where the AI would evaluate the board from the wrong side, leading
to pessimizing strategies since a player want's the worst position for
the opponent. Perhaps here the board evalutation side is only wrong for
every other recursion depth or somthing equally mysterious.

Must investigate.

Also, this commit message is, perhaps, too long.