New Look Ahead Gene functionality
The Look Ahead Gene now has two parameters that can be mutated:
 - mean game length: the expected number of moves in a game for one player
 - positions per second: the number of positions that can be examined per second

When calculating the number of moves to examine, the gene assumes and
Poisson distribution in the total number of moves in a game and
estimates the number of moves left. Using this time and the positions
per second parameter, it determines how many future positions to
examine.