More efficienct return up game tree at end of Look Ahead
After genome.evaluate() is called on the board farthest down the game
tree as Look Ahead Gene will allow, the score and perspective from which
the score was calculated are passed back up the call stack. This allows
the proper perspective to be used at each level of the game tree and
saves on calls to genome.evaluate() since from the opposite perspective
the score is reversed in sign.