Consistently handle stalemate scoring
First attempt at a fix for issue #15.  Needs testing with the players
mentioned in the issue. Needs testing with the AIs mentioned in the
issue.

The main thrust of this fix is to assign a stalemate a score of NaN so
that extra logic in the choose_move()/get_leaf_score() will always
prefer a stalemate to a loss, but without preferring a stalemate to any
other result. Since neither player wants a stalemate, a numeric value
doesn't work since it reverses sign as it goes back up the game tree
several times. NaN allows for a separate valule so that
Genome::evaluate() once again returns opposite values for BLACK vs.
WHITE perspectives.