Reintroduce Branch Pruning Gene
Now that the search_game_tree() algorithm can redistribute the positions it examines, cutting off the search of a game tree branch actually allows other branches to be explored in more depth.

The search_game_tree() method now looks over all the moves without recursing first and picks out the all the moves that look interesting. Then, it recurses through the interesting moves, redistributing positions to examine as it goes.