Make Board::king_is_in_check() non-recursive
Algorithm changed so that each type of piece move is checked to see if it can reach the king's square. This makes copying the board to see if the king is in check after a move unnecessary.

Generating all legal moves now takes 1/3 the time.

Other changes:

Replace pgn_symbol() checks with is_rook(), etc. methods to make code clear.