Overhaul time management in Genetic AI
Instead of calculating the number of game tree nodes to visit ahead of
time, calculate the amount of time to use to choose a move. Then, for
recursing each legal move at a position, divide the time allotted
equally. This is a better adaptive strategy with less data to keep track
of.

Note for the future: the positions_per_second member of the
Look_Ahead_Gene will be tuned for the computer on which it evolves.
Perhaps this should not be an independently evolved factor and instead
be the result of a calibration process upon construction of the Genome.