More logical order for considering promotions
The vast majority of the time, a queen is the right choice for promotions. However, under rare circumstances, the knight is the right choice to avoid a stalemate. Rook and bishop are last since the queen can substitute for either of them.

This change is mostly to see how the choice of promotion pieces will change when the queen is no longer considered last.