Speculative look-ahead implementation
Previously, one of the steps in determining whether the
search_game_tree() function should recurse to look at future moves was
whether there was enough time to look at every possible move. This was
implemented as a simple yes/no based on the number of future legal moves
and the Look_Ahead_Gene::positions_per_second data member. This lead to
situations where many moves were not recursed into for lack of time, but
the last move could take all the time saved and recurse just that move.
This, in turn, lead to weird interactions with principal variations
since, after recursing all the way down (especially to checkmate) the
variation would be rejected because a pseudo-refuting move by the
opponent would not be recursed to reveal that it too lead to a loss.

So, in order to have more evenly distributed depth in game tree
searches, search_game_tree() will recurse even if there is not enough
time allotted. The more time allotted compared to the time required
makes for a larger probability. This should prevent single moves that
fall at the end of the Board::all_legal_moves() list from taking all the
recursion time and lead to greater use of available time.

The probability of recursion is controlled by the
Look_Ahead_Gene::speculation_constant, which is a real number in [0,1].
Zero corresponds with never speculatively recursing and one with always
recursing. This constant (c) gets translated to an exponent by

x = (1-c)/c.

So, c = 0   ==> x = +infinity
    c = 0.5 ==> x = 1
    c = 1   ==> x = 0

This exponent is used as follows in the
Look_Ahead_Gene::enough_time_to_recurse() function:

return Random::success_probability(std::pow(time_alloted/minimum_time, x));

For x = 1, the resultant probability is directly proportional to the
ratio. For x < 1, the resulting probability is greater (e.g., 0.5^0.5 = 0.7),
for x > 1, the probability is lower (e.g., 0.5^2 = 0.25).

Also, spell "allotted" correctly.