Simplify legal move list creation
Instead of creating a list of all moves and then creating a list of
legal moves by filtering that list, create a list of legal moves
directly.

The Sphere of Influence gene uses the list of illegal moves that still
land inside the board to calculate its score. These moves are kept in
the all_other_moves_cache list, which is created in parallel with the
all_legal_moves_cache list.