Number of moves in a chess game are log-normal
For estimating the number of moves in a chess game, the Poisson
distribution has proven to be a poor fit. The log-normal distribution is
much closer to reality, and it has a non-looping solution for
calculating the number of moves remaining, which should allow for much
faster game tree searches.

Idea inspired by: https://chess.stackexchange.com/a/4899/5819