Undo replacement of map with unordered_map
Using std::unordered_map for storing hash data does not result in a
speedup. This is because--as was discovered in the minimal_copy()
commit--that the slowness comes from copying these structures, not from
their use. In all likelihood, the number of items in these structures is
far too small too make a difference.