New Look Ahead Gene property: capture speculation
A second speculation constant is added and is used when there is a
capture available on the board. It is predicted that this new constant
will end up with a larger value since it is important to work out the
consequences of threatened piece captures. A new plot comparing the two
constants has been created to observe if this is true.

A new Board method, capture_possible(), has been created to facilitate
the above gene, as well as tests to prove correct functioning.

A new Math function, clamp, has also been created for when a value needs
to be constrained between two limits.