Remove permanent value from Castling Possible Gene
This value is never set for the original board since the gene only ever
sees copies of the board with hypothetical moves. This means that the
permanent score will most likely not reflect the actual game played.

Also, a typo was fixed in which 0.9 was returned instead of 0.0.

Fixes #56
Reverts 6d13dc4a50a084af1605902a86a0997844f47200