Fix Castling_Possible_Gene fix
Remembered how the scoring system worked for the castling gene. If
castling has not occured yet, then the score should reflect how close to
castling the board position is (how many pieces between the king and
rook have been moved away). The maximum score is only awarded to boards
which have a castling move in the history.