Give King a Piece Strength entry
The piece strength gene now has an entry for kings. However, it does not
affect the noraalization of the piece strengths for two reasons:

1. The king is never removed from the board. Not normalizing with the
   king keeps the Total Force Gene output in the [0,1] range.

2. Early results indicate a tendency to give the king a very negative
   value. This has weird effects on other pieces if the normalization
   includes the king.

The only gene that uses the king strength entry is the Opponent Pieces
Targeted Gene. The negative value would seem to discourage putting the
opponent's king in check. This is odd, but consistent with what happens
with the Sphere of Influence Gene, which also gives negative values to
attack squares near the king.

One interpretation is that attacking the king itself is useless except
for other effects (forking pieces, restricting move choices, threatening
checkmate, etc.). A leaf node in the game tree that has the king being
attacked gets a low score since the consequences can't be evaluated. A
non-leaf node with an attacked king may be temporarily useful in, for
example, the capture of another piece.