Use single file for puttings notes in plots
Example notes file below:

    ID;Game Number;Notes
    121862;140809;To 30-second games
    139746;160648;-Ofast

The file is a semicolon-separated table as above (including header). ID
refers to the Genetic AIs ID and the Game Number to the Round field in
the PGN file created by the gene pool.

The notes will be added to the legend of relevant plots in association
with the vertical lines drawn.