Calculate value of a centipawn
Used for displaying scores with CECP interfaces. The board used to
calibrate looks like this

k......K
........
........
........
........
........
P.......
........

that is, the score of a board with a single white pawn and kings far
enough away that the King Confinement Gene should give nearly the same
score for both kings. The Total Score Gene should provide most of the
difference.