Cut down on piece-specific methods
Instead of methods in the X_pawn_Y- X_rook_X-style, each of these
methods takes a parameter indicating the relevant piece.

Also, the creation of bitwise representations of the square colors is
for the next commit that changes how
Board::enough_material_to_checkmate() works. It could not be separated
from this commit.