Separate Knight moves in Threat_Iterator
Combining all piece moves into a single for() structure was complex and
slow. Separate out knight moves like in the original Board methods is
twice as fast.

Also, this is the first time I've used goto in a program when it seemed
like the best solution (at least until C++ gets coroutines).