Delete test for impossible discovered check
In order for a pawn about to be captured en passant to be blocking a check on the opposite king, that pawn must have moved into a position to block its own attacking piece. This means that the opposite king was already in check, which would be illegal after the previous move.