New working for King Confinement Gene
These new changes to the King Confinement Gene recognize that not all
restrictions to the king's movement are equally good or bad. If a king
can't move because it is surrounded by friendly pieces, then it is
probably protected from attack. If it can't move because it is
surrounded by squares threatened by the enemy, then it is probably in
trouble.  The score_board() function should now be able to distinguish
between these situations.

The importance of the size of the region where the king is safe (free
squares) is yet to be determined.

Effectiveness to be gauged by more gene pool results.